+++ output/kotlin/models/data_class/LogEntry.kt
@@ -2,7 +2,7 @@ package `data_class`
 
 data class LogEntry(
     val deviceState: DeviceState,
-    val timeFired: Int,
+    val timeFired: Long,
     val values: Map<String, String>,
     val tag: String,
     val stackTrace: String,
