+++ output/kotlin/models/interface/All.kt
@@ -73,7 +73,7 @@ interface DeviceState {
 
 interface LogEntry {
     val deviceState: DeviceState
-    val timeFired: Int
+    val timeFired: Long
     val values: Map<String, String>
     val tag: String
     val stackTrace: String
