+++ output/kotlin/models/overridden/All.kt
@@ -144,7 +144,7 @@ data class DeviceState(
 
 interface ILogEntry {
     val deviceState: DeviceState
-    val timeFired: Int
+    val timeFired: Long
     val values: Map<String, String>
     val tag: String
     val stackTrace: String
@@ -154,7 +154,7 @@ interface ILogEntry {
 
 data class LogEntry(
     override val deviceState: DeviceState,
-    override val timeFired: Int,
+    override val timeFired: Long,
     override val values: Map<String, String>,
     override val tag: String,
     override val stackTrace: String,
