+++ output/kotlin/models/overridden_serializable/All.kt
@@ -154,7 +154,7 @@ data class DeviceState(
 
 interface ILogEntry {
     val deviceState: DeviceState
-    val timeFired: Int
+    val timeFired: Long
     val values: Map<String, String>
     val tag: String
     val stackTrace: String
@@ -165,7 +165,7 @@ interface ILogEntry {
 @Serializable
 data class LogEntry(
     @SerialName("device_state") override val deviceState: DeviceState,
-    @SerialName("time_fired") override val timeFired: Int,
+    @SerialName("time_fired") override val timeFired: Long,
     @SerialName("values") override val values: Map<String, String>,
     @SerialName("tag") override val tag: String,
     @SerialName("stack_trace") override val stackTrace: String,
