+++ output/kotlin/models/overridden_serializable/LogEntry.kt
@@ -2,7 +2,7 @@ package `overridden_serializable`
 
 interface ILogEntry {
     val deviceState: DeviceState
-    val timeFired: Int
+    val timeFired: Long
     val values: Map<String, String>
     val tag: String
     val stackTrace: String
@@ -13,7 +13,7 @@ interface ILogEntry {
 @Serializable
 data class LogEntry(
     @SerialName("device_state") override val deviceState: DeviceState,
-    @SerialName("time_fired") override val timeFired: Int,
+    @SerialName("time_fired") override val timeFired: Long,
     @SerialName("values") override val values: Map<String, String>,
     @SerialName("tag") override val tag: String,
     @SerialName("stack_trace") override val stackTrace: String,
